Update MovieCard to have hover effect
